requirements.txt
# requirements.txt
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.11.0
pytest>=7.0.0